+++ Mercury/Assets/Scripts/Weapons/Projectiles/RPG.cs
@@ -10,10 +10,10 @@ public class RPG : Projectile
         base.Init();
         // Stats
         speed = 10f;
-        damage = 100;
+        damage = 100f;
         blastRadius = 1f;
     }
-    public void setDamage(int damageA)
+    public void setDamage(float damageA)
     {
         damage = damageA;
     }
